requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6  # For handling file uploads
pyyaml>=6.0.1  # For configuration
python-dotenv>=1.0.0  # For environment variables

# OpenAI API (Primary STT)
openai>=1.0.0  # OpenAI Whisper API client
aiohttp>=3.8.0  # For async HTTP requests

# Audio processing
librosa>=0.10.1
pydub>=0.25.1
SoundFile>=0.12.1
ffmpeg-python>=0.2.0

# Voice authentication
resemblyzer>=0.1.1.dev0  # For voice embedding and verification
speechbrain>=0.5.15
torch>=2.1.0  # PyTorch for deep learning
torchaudio>=2.1.0  # Audio processing for PyTorch

# Speech recognition (Fallback)
whisper>=20231117  # Local Whisper as fallback
numba>=0.58.1  # Required for Whisper

# Hybrid STT
sentence-transformers>=0.12.2  # For semantic similarity

# Security
python-jose>=3.3.0  # For JWT
cryptography>=41.0.4  # For encryption
passlib>=1.7.4  # For password hashing

# Utilities
numpy>=1.24.4
requests>=2.31.0
tenacity>=8.2.3  # For retry mechanisms
tqdm>=4.66.1  # For progress bars
pyaudio>=0.2.13  # For audio recording

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.0  # For testing FastAPI
black>=23.10.1  # For code formatting
isort>=5.12.0  # For import sorting
mypy>=1.6.1  # For type checking
